moved onedrive platform files to .net standard project
